plot robustness; handle coordinates files with and without FIBER column
